dry CloseIconMini
