+++ src/components/AlertBox.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import Text from './Text'
 import FlexRow from './FlexRow'
 import DOMPurify from 'dompurify'
 import MegaphoneIcon from '../icons/Megaphone'
+import { CloseIconMini } from '../icons/Close'
 
 interface AlertBoxProps {
   children: React.ReactNode
@@ -11,7 +11,7 @@ interface AlertBoxProps {
   onDismiss?: () => void
 }
 
-export function AlertBox({ children, icon, onClick, onDismiss }: AlertBoxProps) {
+function AlertBox({ children, icon, onClick, onDismiss }: AlertBoxProps) {
   const style: Record<string, React.CSSProperties> = {
     alert: {
       backgroundColor: 'var(--purple20)',
@@ -37,18 +37,10 @@ export function AlertBox({ children, icon, onClick, onDismiss }: AlertBoxProps) 
             <div style={style.icon}>{icon}</div>
             {children}
           </FlexRow>
-          {onDismiss ? <CloseIcon onClick={onDismiss} /> : null}
+          {onDismiss ? <CloseIconMini onClick={onDismiss} /> : null}
         </FlexRow>
       </div>
     </div>
-  )
-}
-
-export function AlertText({ children }: { children: React.ReactNode }) {
-  return (
-    <Text color='black' bold smaller wrap>
-      {children}
-    </Text>
   )
 }
 
@@ -64,26 +56,5 @@ export function InfoBox({ html }: { html: string }) {
     <AlertBox icon={<MegaphoneIcon animated />}>
       <div dangerouslySetInnerHTML={{ __html: sanitizedHtml }} style={style} />
     </AlertBox>
-  )
-}
-
-function CloseIcon({ onClick }: { onClick: () => void }) {
-  const style = {
-    cursor: 'pointer',
-  }
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width='20'
-      height='20'
-      viewBox='0 0 512 512'
-      onClick={onClick}
-      style={style}
-    >
-      <path
-        fill='currentColor'
-        d='M256 48C141.31 48 48 141.31 48 256s93.31 208 208 208s208-93.31 208-208S370.69 48 256 48Zm75.31 260.69a16 16 0 1 1-22.62 22.62L256 278.63l-52.69 52.68a16 16 0 0 1-22.62-22.62L233.37 256l-52.68-52.69a16 16 0 0 1 22.62-22.62L256 233.37l52.69-52.68a16 16 0 0 1 22.62 22.62L278.63 256Z'
-      />
-    </svg>
   )
 }
